No dependencies on the old NetcdfC++ interface MAGP-863

+++ src/decoders/NetcdfOrcaInterpretor.cc
@@ -43,7 +43,7 @@ bool NetcdfOrcaInterpretor::interpretAsMatrix(Matrix** data)
 
 	Netcdf netcdf(path_, dimension_method_);
 	NetVariable var = netcdf.getVariable(longitude_);
-	vector<long> dims;
+	vector<size_t> dims;
 	var.getDimensions(dims);
